slibtool-shared: properly handle .la arguments when creating an archive.
tool
====
slibtool-shared (or slibtool with -disable-static).

scenario
========
- create libfoo.so (via libfoo.la).
- create libbar.a  (via libbar.la), with libfoo.la as an input argument.

Before this fix, slibtool failed since it could not find libfoo.a.slibtool.deps.

After this fix, slibtool properly handles this scenario, first by detecting that
there is no archive for which secondary dependencies need to be processed, and then
by detecting that libfoo.a is a symlink to /dev/null.