build system: PE, ELF: defined DSO_REF_VER, DSO_REF_SONAME, DSO_REF_SOLINK.
These definitions come in handy when a project generates alongside
its ``primary shared library'' one or more ``extension libraries''
that depend on it. When the rules for generating extension
libraries use the $^ directive, the above dependency must
be declared in a target-aware manner.

In most cases, one would want to express this dependency by way
of $(DSO_REF_SONAME), thereby pulling in lib/libfoo.so.$(MAJOR)
on ELF targets, lib/libfoo.$(MAJOR).lib.a on midipix targets,
and lib/libfoo.$(MAJOR).dll.a on win32 targets.